requirements.txt
asgiref==3.8.1
Django==5.0.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
flake8==7.1.1
flake8-html==0.4.3
Jinja2==3.1.4
MarkupSafe==2.1.5
mccabe==0.7.0
pycodestyle==2.12.1
pyflakes==3.2.0
Pygments==2.18.0
PyJWT==2.9.0
sqlparse==0.5.1
tzdata==2024.1
